Annual Summary of Hatchery

Summary of collections and releases from the Klamath Falls National Fish Hatchery since its inception in 2016. Data summarizes the annual inputs and outputs of the hatchery program by federal fiscal year Corresponds to **Table 1** and **Table 4** in the FY2024 KFNFH Annual Report.

KFNFH_annual_summary

Format

A tibble with 9 rows and 12 columns

hatchery_name: name of hatchery (KFNFH)

fiscal_year: fiscal year

larvae_collected: number of wild larval suckers collected

sarp_release: number of fish released or transferred that were raised in the Sucker Assisted Rearing Program (SARP)

sarp_tl_mm: average total length (TL) of the SARP release fish, measured in millimeters (mm)

fingerling_release: number of fingerlings released or transferred

fingerling_tl_mm: average total length (TL) of the fingerling release fish, measured in millimeters (mm)

fry_release: number of fry released or transferred. This includes fry produced from experimental, wild-spawning activities that were transferred to other facilities (e.g., the Klamath Tribes Hatchery in FY24)

fry_tl_mm : average total length (TL) of the fry release fish, measured in millimeters (mm)

salvage_release: number of suckers salvaged throughout the year

salvage_sl_mm: average length of release fish, measured in millimeters (mm)

salvage_tl_mm: standard lengths (SL) of salvage fish are averages from measurements for each year

primary_collection_method: drift net, dip net, or both

Source

Klamath Falls National Fish Hatchery Annual Report for Fiscal Year 2024