Data Location Lookup

This dataset compiles location data relevant to fisheries data collection efforts from across the Klamath Basin. This dataset is used to map data collection locations.

data_location_lookup

Format

A tibble with 35 rows and 10 columns

stream: stream

sub_basin: sub-basin name (upper klamath, lower klamath, trinity, shasta)

data_type: type of data (rst, hatchery, redd/carcass survey)

site_name: name of site such as a RST site or hatchery location (big bar, shasta river, bogus, willow creek, pear creek, weitchpec, iron gate fish hatchery, trinity river hatchery, klamath hatchery)

agency: agency that manages/monitors the site (arcata fwo, arcata fwo, cdfw, hoopa tribal fisheries department, karuk, yurok tribal fisheries program, usfws)

latitude: longitude

longitude: longitude

downstream_latitude: Latitude of the downstream end of the feature's extent, applicable for survey extents

downstream_longitude: Longitude of the downstream end of the feature's extent, applicable for survey extents

link: web link containing more information about fisheries locations