Salmon Spawner Escapement

Digital version of [CDFW's Spring Chinook Salmon Megatable](https://wildlife.ca.gov/Conservation/Fishes/Chinook-Salmon/Anadromous-Assessment) and [CDFW's Spring Chinook Salmon Megatable](https://casalmon.org/wp-content/uploads/2024/08/FINAL-2023-Spring-Chinook-Megatable-v.28-Mar-2024.pdf) Spawner Escapement sections as well as other data sources for Coho and Steelhead not included in the Megatable.

salmon_spawner_escapement

Format

A tibble with 1,722 rows and 6 columns

location: River, hatchery, or basin where escapement is reported. ("bogus creek", "hoopa and yurok tributaries", "hoopa trinity tributaries", "iron gate hatchery", "klamath basin", "klamath river", "lower klamath river", "other klamath tributaries", "other klamath trinity tributaries", "other trinity tributaries", "salmon river", "scott river", "shasta river", "south fork trinity river", "trinity basin", "trinity river", "trinity river hatchery", "yurok klamath tributaries")

year: Return year of the escapement estimate/count (1978-2024)

species: Species name (spring chinook salmon, fall chinook salmon, coho salmon, winter steelhead, steelhead)

origin: Fish origin category (hatchery, wild, mixed, unknown)

lifestage: Life stage reported (adult, adult and subadult, grilse). In a few cases adult and subadult were not distinguished and are grouped together.

estimate_type: Type of estimate reported (abundance, count, redd abundance). Work in progress as these have not been filled in yet for the data from the Megatables.

estimate: value of estimate

confidence_interval: type of confidence or credible interval (e.g., 95)

lower_bounds_estimate: value of lower bounds estimate, if applicable

upper_bounds_estimate: value of upper bounds estimate, if applicable

estimation_method: type of model used (count, net upstream count, upstream count, expanded redd estimate, mark recapture. Work in progress as these have not been filled in yet for the data from the Megatables.)

is_complete_estimate: data included in model are complete. In some cases, it is known that an estimate is not complete, otherwise it is assumed that estimate is complete.

source: describes where data were sourced from. Currently there are 2 sources: (1) California Department of Fish and Wildlife (CDFW) document library (https://www.nrm.dfg.ca.gov/documents/ContextDocs.aspx?cat=Fisheries–AnadromousSalmonidPopulationMonitoring), (2) CDFW Spring and Fall Megatables