Salmon Juvenile Abundance

This dataset compiles modeled estimates and credible intervals from fisheries models such as abundance and survival estimates. This dataset is not species or location specific. This is a key dataset! In the future, this dataset may be broken down into separate tables like what was done for spawner escapement.

salmon_juvenile_abundance

Format

A tibble with 122 rows and 14 columns

julian_year: julian year (2001-2022). Min/max years vary by location, species, type of estimate. Juvenile abundance typically spans multiple years (e.g. Oct 2020 - May 2021) and would be assigned a julian year for the latter part of this period.

stream: stream ("scott river", "shasta river")

species: species (coho salmon, steelhead, fall chinook salmon)

origin: origin (wild)

lifestage: lifestage category (yoy, age 1+, age 2+, smolt)

sex: sex, if applicable to model (NA)

estimate_type: type of model estimate (abundance, count)

estimate: value of estimate

confidence_interval: type of confidence or credible interval (e.g., 95)

lower_bounds_estimate: value of lower bounds estimate, if applicable

upper_bounds_estimate: value of upper bounds estimate, if applicable

estimation_method: type of model used (mark-recapture)

is_complete_estimate: data included in model are complete. In some cases, it is known that an estimate is not complete, otherwise it is assumed that estimate is complete.

source: describes where data were sourced from. Currently there is one data source: (1) California Department of Fish and Wildlife (CDFW) document library (https://www.nrm.dfg.ca.gov/documents/ContextDocs.aspx?cat=Fisheries–AnadromousSalmonidPopulationMonitoring)